feat: indexとsurveyの統合
